+++ src/components/Counter.js
@@ -1,0 +1,45 @@
+import React, { useState } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import {
+  decrement,
+  increment,
+  incrementByAmount,
+  incrementAsync,
+  incrementIfOdd,
+  selectCount,
+} from '../redux/counter/counterSlice';
+
+
+export function Counter() {
+  const count = useSelector(selectCount);
+  const dispatch = useDispatch();
+  const [incrementAmount, setIncrementAmount] = useState('2');
+
+  const incrementValue = Number(incrementAmount) || 0;
+
+  return (
+    <div>
+      <div>
+        <button onClick={() => dispatch(decrement())}>
+          -
+        </button>
+        <span>{count}</span>
+        <button onClick={() => dispatch(increment())}>
+          +
+        </button>
+      </div>
+      <div>
+        <input value={incrementAmount} onChange={(e) => setIncrementAmount(e.target.value)}/>
+        <button onClick={() => dispatch(incrementByAmount(incrementValue))}>
+          Add Amount
+        </button>
+        <button onClick={() => dispatch(incrementAsync(incrementValue))}>
+          Add Async
+        </button>
+        <button onClick={() => dispatch(incrementIfOdd(incrementValue))}>
+          Add If Odd
+        </button>
+      </div>
+    </div>
+  );
+}
